+++ deltafi-core/src/main/java/org/deltafi/core/DeltaFiCoreApplication.java
@@ -19,12 +19,14 @@ package org.deltafi.core;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableAsync;
 
 @SpringBootApplication
 @EnableAsync
 @EnableRetry
+@EnableCaching
 public class DeltaFiCoreApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(DeltaFiCoreApplication.class, args);